refactor(CLI): Rename load-dotenv util
