+++ lib/Serverless.js
@@ -25,7 +25,7 @@ const resolveLocalServerlessPath = require('./cli/resolve-local-serverless-path'
 const isHelpRequest = require('./cli/is-help-request');
 const resolveCliInput = require('./cli/resolve-input');
 const readConfiguration = require('./configuration/read');
-const loadEnv = require('./loadEnv');
+const loadDotenv = require('./cli/load-dotenv');
 
 class Serverless {
   constructor(config) {
@@ -226,7 +226,7 @@ class Serverless {
       _.get(configurationInput, 'provider.stage', 'dev');
 
     if (configurationInput.useDotenv) {
-      await loadEnv(stage);
+      await loadDotenv(stage);
     } else {
       const defaultEnvFilePath = path.join(process.cwd(), '.env');
       const stageEnvFilePath = path.join(process.cwd(), `.env.${stage}`);
